moved motion energy to discovery step
