+++ asoid/apps/F_view.py
@@ -2,7 +2,7 @@ import categories
 import streamlit as st
 from app import swap_app
 from config.help_messages import IMPRESS_TEXT
-from utils.view_results import Viewer, MotionEnergyMachine
+from utils.view_results import Viewer
 
 
 CATEGORY = categories.VIEW
@@ -13,11 +13,6 @@ def main(config=None):
     st.subheader("Viewer")
     viewer = Viewer(config)
     viewer.main()
-
-    if config is not None:
-        st.subheader("Motion Energy")
-        motion_energy = MotionEnergyMachine(config)
-        motion_energy.main()
 
     bottom_cont = st.container()
     with bottom_cont:
